requirements.txt
# Python dependencies for Strava Dashboard
Flask==2.2.3
Flask-Session==0.4.0
requests==2.28.2
python-dotenv==1.0.0
gunicorn==20.1.0
